Add files via upload
At opening, a security warning should display "External data connections have been disabled".
This is normal as the file has been created with two connections that query Coingecko to obtain the price of XIO and Flash.

To verify each connections, you can go to the Data tab, then click on "Connections". 

To check the details of each connections:
Go to the sheet "Connections", then click on one of the table. 
Click on the Query tab, then "Edit" button, it will open a new window.
Double-click on "Source" in "Applied steps" list on the right of the window to display the data source (CoinGecko URL)